Cloud Functions/api/requirements.txt
functions-framework==3.*
google-cloud-firestore
flask
